refactor: Enhance navigation calculations and streamline task execution logic

+++ src/main/java/com/devone/bot/core/brain/navigator/tags/BotFovSliceTagger.java
@@ -15,7 +15,7 @@ public class BotFovSliceTagger {
      */
     public static int tagFovSliceAll(List<BotBlockData> blocks,
                                      BotPositionSight bot,
-                                     float fovDeg,
+                                     double fovDeg,
                                      int radius,
                                      int height) {
